scripts/srd_parser/requirements.txt
pdfplumber>=0.10.0
python-slugify>=8.0.0
